[SharedGC] Register SharedCC stub switch callback in NAPI layer
Register ExecuteSharedCCStubSwitchCallback in ArkNativeEngine
constructor, following the same pattern as SetTriggerGCTaskCallback.
PostSharedCCStubSwitchTask uses EventHandler::PostTask to dispatch
stub switch execution to the main thread event loop.

This removes the runtime's direct dependency on UV loop for SharedCC
main thread stub switching.

Key changes:
- JSNApi::SetExecuteSharedCCStubSwitchCallback registration in constructor
- PostSharedCCStubSwitchTask method using EventHandler::PostTask

Issue: openharmony/arkcompiler_ets_runtime#13292

+++ native_engine/impl/ark/ark_idle_monitor.cpp
@@ -795,8 +795,8 @@ ArkIdleMonitor::WorkerGCResult ArkIdleMonitor::EvaluateWorkerGC(napi_env workerE
 void ArkIdleMonitor::PostWorkerFullGC(napi_env workerEnv, bool notifyFinished)
 {
     auto arkNativeEngine = reinterpret_cast<ArkNativeEngine*>(workerEnv);
-    std::pair<void*, uint8_t> data(reinterpret_cast<void*>(const_cast<EcmaVM*>(arkNativeEngine->GetEcmaVm())),
-                                   static_cast<uint8_t>(TRIGGER_IDLE_GC_TYPE::FULL_GC));
+    std::pair<void*, uint16_t> data(reinterpret_cast<void*>(const_cast<EcmaVM*>(arkNativeEngine->GetEcmaVm())),
+                                   static_cast<uint16_t>(TRIGGER_IDLE_GC_TYPE::FULL_GC));
     auto callbackTask = [this, workerEnv, notifyFinished]() {
         HILOG_DEBUG("ArkIdleMonitor: try trigger thread full gc end");
         UnRegisterSentTaskWorkerEnv(workerEnv);
